+++ common/src/main/java/com/evandev/fieldguide/client/gui/screens/FieldGuideJournalScreen.java
@@ -51,7 +51,7 @@ public class FieldGuideJournalScreen extends BookScreen {
         int titleY = this.leftPageBounds.top() + 8;
         int textY = titleY + this.font.lineHeight * 2 + 13;
         int textAreaWidth = this.rightPageBounds.width() - 12;
-        int textAreaHeight = this.leftPageBounds.height() - (textY - this.leftPageBounds.top()) - 20;
+        int textAreaHeight = this.leftPageBounds.height() - (textY - this.leftPageBounds.top()) - 11;
 
         // Journal Title
         if (currentSpread == 0) {